Update rcon.py
Initialise the protocol factory with rcon._loop, rather than the passed loop variable. If the user does not pass a loop in the initialisation, loop will still be null however rcon._loop is initialised with the asyncio.get_event_loop. Therefore, use the rcon._loop in the protocol factory's initialisation.